"internal: Classes moved to proper packages to increase readability"
