+++ app/src/main/java/com/org/tmdb/ui/screens/trending/TrendingList.kt
@@ -1,36 +1,28 @@
-package com.org.tmdb.ui.viewmodels
+package com.org.tmdb.ui.screens.trending
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.progressSemantics
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.CircularProgressIndicator
-import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.State
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import coil.compose.AsyncImage
@@ -38,12 +30,13 @@ import com.org.tmdb.BuildConfig
 import com.org.tmdb.data.remote.MediaType
 import com.org.tmdb.data.remote.ResultTrending
 import com.org.tmdb.data.remote.TimeWindow
-import com.org.tmdb.data.remote.Trending
 import com.org.tmdb.ui.common.CommonSpacer
 import com.org.tmdb.ui.preview.TrendingParameterProvider
 import com.org.tmdb.ui.theme.primaryDarkMode
 import com.org.tmdb.ui.theme.primaryLightMode
 import com.org.tmdb.ui.theme.white
+import com.org.tmdb.ui.viewmodels.MainActivityUiState
+import com.org.tmdb.ui.viewmodels.MainViewModel
 
 
 @Composable
